don't use typing with ServerDefaultType (#180)
The server default is any expression that is passed as DDL
to the database and these usually don't have SQL types explicitly
stated.

With the code as is, pylance is complaining about this:

    Column(DateTime(), server_default=func.now())

and requiring I do this:

    Column(DateTime(), server_default=func.now(type_=DateTime))

people don't need to do that, server_defaults SQL type always comes
from the column type and doesn't normally need to be stated.

Also, column_server_default.py was importing "functions as func",
which is wrong. "func" is not a module it's a namespace object,
fixed that.